{"schema":"decodex/commit/1","summary":"Split status execution program readback","authority":"XY-1115"}
